requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
imageio==2.36.0
imutils==0.5.4
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
tifffile==2024.9.20
